Implement serialization for extension fields
Extension fields can be added to the `StatusBuilder` using the
`add_extension` method.

The map type was switched to a `BTreeMap` in order to be able to sort
the keys alphabetically when serializing.

I also changed the struct serialization from using `serialize_struct`
to `serialize_map`. Reason: https://github.com/serde-rs/serde/issues/708
In our case we only serialize to JSON anyways, so there's no need for
serialization formats that can differentiate between structs and maps.